refactor(vite-app): refactor fields
