+++ packages/vite-app/src/components/class/index.ts
@@ -1,1 +1,3 @@
 export * from "./ClassView";
+export * from "./Header";
+export * from "./Fields";
