--- packages/vite-app/src/components/fields/index.ts
@@ -1,1 +1,0 @@
-export * from "./Fields";
